+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -1,21 +1,14 @@
 package org.apache.coyote.http11;
 
-import static org.apache.coyote.HttpHeaders.CONTENT_LENGTH;
-
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.net.Socket;
-import java.util.ArrayList;
-import java.util.List;
 import nextstep.jwp.presentation.Controller;
-import org.apache.coyote.HttpBody;
-import org.apache.coyote.HttpHeaders;
 import org.apache.coyote.HttpRequest;
 import org.apache.coyote.HttpResponse;
-import org.apache.coyote.HttpStartLine;
 import org.apache.coyote.Processor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -23,9 +16,6 @@ import org.slf4j.LoggerFactory;
 public class Http11Processor implements Runnable, Processor {
 
     private static final Logger LOG = LoggerFactory.getLogger(Http11Processor.class);
-
-    private static final String LINE_BEFORE_READ = " ";
-    private static final int START_LINE = 0;
 
     private final Socket connection;
 
@@ -44,55 +34,14 @@ public class Http11Processor implements Runnable, Processor {
              final BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(inputStream));
              final OutputStream outputStream = connection.getOutputStream()) {
 
-            final HttpRequest httpRequest = readHttpRequest(bufferedReader);
+            final HttpRequest httpRequest = RequestParser.readHttpRequest(bufferedReader);
             final HttpResponse httpResponse = new HttpResponse();
 
             doService(httpRequest, httpResponse);
-
             write(outputStream, httpResponse);
         } catch (final Exception e) {
             LOG.error(e.getMessage(), e);
         }
-    }
-
-    private HttpRequest readHttpRequest(final BufferedReader bufferedReader) throws IOException {
-        final List<String> rawHeaders = readHeaders(bufferedReader);
-
-        final String rawStartLine = rawHeaders.remove(START_LINE);
-        final HttpStartLine startLine = HttpStartLine.from(rawStartLine);
-        final HttpHeaders headers = HttpHeaders.from(rawHeaders);
-        final HttpBody body = HttpBody.from(readBody(bufferedReader, headers));
-
-        LOG.info("============= HTTP REQUEST =============");
-        LOG.info("{}\n{}\n\n{}", rawStartLine, headers, body);
-
-        return new HttpRequest(startLine, headers, body);
-    }
-
-    private static List<String> readHeaders(final BufferedReader bufferedReader) throws IOException {
-        final List<String> rawHttpRequest = new ArrayList<>();
-
-        String line = LINE_BEFORE_READ;
-        while (!line.isEmpty()) {
-            line = bufferedReader.readLine();
-            rawHttpRequest.add(line);
-        }
-        return rawHttpRequest;
-    }
-
-    private static String readBody(final BufferedReader bufferedReader, final HttpHeaders headers) throws IOException {
-        final String header = headers.getHeader(CONTENT_LENGTH);
-        if (header == null) {
-            return null;
-        }
-
-        final int contentLength = Integer.parseInt(header);
-        final char[] buffer = new char[contentLength];
-        if (bufferedReader.ready()) {
-            bufferedReader.read(buffer);
-        }
-
-        return new String(buffer);
     }
 
     private void doService(final HttpRequest request, final HttpResponse response) throws Exception {
